requirements.txt
pandas
numpy
fastapi
uvicorn
python-multipart
pillow
tensorflow==2.18.0